+++ src/frontend/web/src/pages/CreateWs.tsx
